+++ src/eps/impls/krylov/krylovschur/ks-indef.c
@@ -100,7 +100,7 @@ PetscErrorCode EPSSolve_KrylovSchur_Indefinite(EPS eps)
     if (l) { ierr = PetscInfo1(eps,"Preparing to restart keeping l=%" PetscInt_FMT " vectors\n",l);CHKERRQ(ierr); }
 
     if (eps->reason == EPS_CONVERGED_ITERATING) {
-      if (breakdown) SETERRQ1(PetscObjectComm((PetscObject)eps),PETSC_ERR_CONV_FAILED,"Breakdown in Indefinite Krylov-Schur (beta=%g)",beta);
+      if (breakdown) SETERRQ1(PetscObjectComm((PetscObject)eps),PETSC_ERR_CONV_FAILED,"Breakdown in Indefinite Krylov-Schur (beta=%g)",(double)beta);
       else {
         /* Prepare the Rayleigh quotient for restart */
         ierr = DSTruncate(eps->ds,k+l,PETSC_FALSE);CHKERRQ(ierr);
