+++ src/sys/classes/st/impls/filter/filter.c
@@ -398,7 +398,7 @@ PetscErrorCode STView_Filter(ST st,PetscViewer viewer)
     ierr = PetscViewerASCIIPrintf(viewer,"  Filter: numerical range is [%g,%g]\n",(double)ctx->left,(double)ctx->right);CHKERRQ(ierr);
     ierr = PetscViewerASCIIPrintf(viewer,"  Filter: degree of filter polynomial is %" PetscInt_FMT "\n",ctx->polyDegree);CHKERRQ(ierr);
     if (st->state>=ST_STATE_SETUP) {
-      ierr = PetscViewerASCIIPrintf(viewer,"  Filter: limit to accept eigenvalues: theta=%g\n",ctx->filterInfo->yLimit);CHKERRQ(ierr);
+      ierr = PetscViewerASCIIPrintf(viewer,"  Filter: limit to accept eigenvalues: theta=%g\n",(double)ctx->filterInfo->yLimit);CHKERRQ(ierr);
     }
   }
   PetscFunctionReturn(0);
